+++ demos/react-native-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,15 @@
 # react-native-supabase-todolist
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [3aaee03]
+  - @journeyapps/powersync-sdk-common@1.6.0
+  - @journeyapps/powersync-attachments@2.0.3
+  - @journeyapps/powersync-react@1.1.3
+  - @journeyapps/powersync-sdk-react-native@1.4.2
+
 ## 1.1.2
 
 ### Patch Changes